+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- updater: handling of "reboot is required" from updater supervisor;
+    it is now possible to approve update request in reForis
+
 ## [5.1.0] - 2022-06-27
 ### Fixed
 - networks: Sort interfaces by port names and disregard order in uci config